Add class and ID for elements
Adds an ID for the second element to give it blue text
and a font size of 36px.

Creates a class for adding a font size of 24px which needs to go
to the third and fourth element.

Adds an ID to the fourth element to give it a background
color of light green and a bold font.